packages/nest-api/src/auth-module/custom-validation/unique-email.validation.ts
import {
  ValidatorConstraint,
  ValidatorConstraintInterface,
} from 'class-validator';
import { Injectable } from '@nestjs/common';
import { UserService } from '../services/user.service';

@ValidatorConstraint({ name: 'uniqueEmail', async: true })
@Injectable()
export class UniqueEmailValidation implements ValidatorConstraintInterface {
  constructor(private readonly userService: UserService) {}

  async validate(value: string): Promise<boolean> {
    return !(await this.userService.findOneByEmail(value));
  }

  defaultMessage() {
    return `Email already exist`;
  }
}
